+++ client/src/components/Dashboard/BlockedDomains.js
@@ -58,7 +58,7 @@ const BlockedDomains = ({
                 noDataText={t('no_domains_found')}
                 minRows={6}
                 defaultPageSize={100}
-                className="-highlight card-table-overflow stats__table"
+                className="-highlight card-table-overflow--limited stats__table"
             />
         </Card>
     );