+++ client/src/components/Dashboard/QueriedDomains.js
@@ -59,7 +59,7 @@ const QueriedDomains = ({
             noDataText={t('no_domains_found')}
             minRows={6}
             defaultPageSize={100}
-            className="-highlight card-table-overflow stats__table"
+            className="-highlight card-table-overflow--limited stats__table"
         />
     </Card>
 );