+++ client/src/components/Filters/Form.js
@@ -28,6 +28,7 @@ const Form = (props) => {
                         className="form-control"
                         placeholder={t('enter_name_hint')}
                         validate={[required]}
+                        normalizeOnBlur={data => data.trim()}
                     />
                 </div>
                 <div className="form__group">
@@ -39,6 +40,7 @@ const Form = (props) => {
                         className="form-control"
                         placeholder={t('enter_url_or_path_hint')}
                         validate={[required, isValidPath]}
+                        normalizeOnBlur={data => data.trim()}
                     />
                 </div>
                 <div className="form__description">